refactor: move summarize feature to a specific folder
